routeProject/requirements.txt
pandas>=1.3.0
requests>=2.26.0
ortools>=9.0.0
geopy>=2.2.0
matplotlib>=3.5.0
pillow>=9.0.0